[mvsUtils] Rename depthMapIO to mapIO
