+++ src/aliceVision/mvsUtils/mapIO.cpp
@@ -4,7 +4,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "depthSimMapIO.hpp"
+#include "mapIO.hpp"
 
 #include <aliceVision/mvsData/Point2d.hpp>
 #include <aliceVision/mvsUtils/fileIO.hpp>
