+++ src/aliceVision/mvsUtils/mapIO.hpp
